Crud

+++ src/components/post/PostCard.js
@@ -1,4 +1,5 @@
-  import React from 'react';
+import React from 'react';
+import { Link } from 'react-router-dom';
 import { deletePost, getPosts } from './PostManager';
  
   
@@ -19,28 +20,39 @@ import { deletePost, getPosts } from './PostManager';
       };
     const sessionUserId = localStorage.getItem("rare_userid")
     const userid =  parseInt(sessionUserId)
-    const handleDelete= id => {
-      deletePost(id)
-      .then(() => getPosts().then(setPosts));
+    const id = post.id
+    const handleMethod= (method,id) => {
+      if (method === 'delete') {
+        deletePost(id)
+        .then(() => getPosts().then(setPosts));
+      } else if  (method === 'edit'){
+        console.warn('test edit')
+      } else if  (method === 'detail'){
+        console.warn('test detail')
+      }
     };
     return (
       <div>
         <div className="card" style={{ width: '18rem', margin: '3px' }}>
           <div className="card-body">
             <h1 className="card-title">{post.title}</h1>
-            <h5 className="card-title">{post.content}</h5>
-            <h5 className="card-title">{post.publication_date}</h5>
+            {/* <h5 className="card-title">{post.content}</h5> */}
+            {/* <h5 className="card-title">{post.publication_date}</h5> */}
             <h5 className="card-title">{renderCategory(post.category_id)}</h5> 
-            <button type="button"> Detail </button>    
+            <Link to={`/detail/${id}`} >Detail</Link>   
+            {/* <button type="button"  onClick={() => handleMethod('detail',post.id)}> Detail </button>     */}
             { userid === post.user_id ?  (
-                <button type="button"> Edit </button>
+                // <button type="button"  onClick={() => handleMethod('edit',post.id)} > Edit </button>
+                <Link to={`/edit/${id}`} >
+                Edit
+              </Link>
                 
             ) : (
             ''    
             )
             } 
             { userid === post.user_id ?  (
-                <button type="button" Delete onClick={() => handleDelete(post.id)} > Delete</button>                
+                <button type="button"  onClick={() => handleMethod('delete',post.id)} > Delete</button>                
             ) : (
             ''    
             )
